Properly handle front end init failure
It was possible for the client to not exit if an error occured during front end initialization that would also prevent the usage of _Front::postError().
This would cause an exception to go all the way up the stack and prevent stop() from being called.

We only change the frontend (from the MiniUI) to the real one after _Front::start() is called and we protected ourselves against exceptions in _Front::portError()